+++ ingraind-probes/src/connection/mod.rs
@@ -46,6 +46,6 @@ pub struct Connection {
 
 #[derive(Debug)]
 pub enum Message {
-    Send(Connection, u16),
-    Receive(Connection, u16)
+    Send(Connection, u64),
+    Receive(Connection, u64)
 }
